kernel: net: increase sk_[max_]ack_backlog.

+++ linux-3.4.x/include/net/sock.h
@@ -339,8 +339,8 @@ struct sock {
 	rwlock_t		sk_callback_lock;
 	int			sk_err,
 				sk_err_soft;
-	unsigned short		sk_ack_backlog;
-	unsigned short		sk_max_ack_backlog;
+	u32			sk_ack_backlog;
+	u32			sk_max_ack_backlog;
 	__u32			sk_priority;
 #ifdef CONFIG_CGROUPS
 	__u32			sk_cgrp_prioidx;
